requirements.txt
ansible
argh
argcomplete
pychef
requests
